requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
pymysql==1.1.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
fake-useragent==1.4.0
gunicorn==21.2.0
cryptography==41.0.7
APScheduler==3.10.4
Flask-Mail==0.9.1
openpyxl==3.1.2
pandas==2.2.2
python-dateutil==2.9.0
